pyproject.toml
[tool.poetry]
name = "acr-ingest"
version = "0.0.0"
description = "Ingest daily ACRCloud data dumps"
authors = ["RaBe IT-Reaktion <it@rabe.ch>"]
license = "AGPLv3"
readme = "README.md"
packages = [{include = "acringest.py"}]

[tool.poetry.scripts]
acringest = 'acringest:main'

[tool.poetry.dependencies]
python = "^3.11"
cloudevents = "^1.9.0"
ConfigArgParse = "^1.5.3"
kafka-python = "^2.0.2"
minio = "^7.1.15"
jsondiff = "^2.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
pytest-random-order = "^1.1.0"
pytest-ruff = "^0.0.6"
ruff = "^0.0.270"
isort = "^5.12.0"
pytest-isort = "^3.1.0"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "7.2"
addopts = "-ra -q --random-order --doctest-glob='*.md' --doctest-modules --cov=acringest --cov-fail-under=100 --ruff --isort"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
